[Rename] 파일 위치 변경
ref : #93
