+++ backend/src/main/java/site/bookmore/bookmore/oauth2/CustomOAuth2UserService.java
@@ -1,4 +1,4 @@
-package site.bookmore.bookmore.oauth2.service;
+package site.bookmore.bookmore.oauth2;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
